+++ NEWS.md
@@ -1,9 +1,12 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
-# shinywqbench 1.0.0
+# shinywqbench (development version)
+
+- Updating title, adding instructions to data tab, updating user and about tabs
+
+# shinywqbench 0.1.0
 
 - App is fully functional
-
 
 # shinywqbench 0.0.0.9000